fix ordering of addFormats. New ajvOptions param

+++ tests/src/format-schema.test.ts
@@ -1,7 +1,7 @@
 import path from "path";
 import fs from "fs";
 import { generate } from "openapi-typescript-validator";
-import Ajv from 'ajv';
+import Ajv from "ajv";
 
 describe("format-schema - compile based", () => {
   const name = "format";
@@ -9,8 +9,7 @@ describe("format-schema - compile based", () => {
   const schemaDir = path.join(__dirname, "../schemas");
 
   beforeAll(async () => {
-    if (fs.existsSync(generatedDir))
-      fs.rmdirSync(generatedDir, { recursive: true });
+    if (fs.existsSync(generatedDir)) fs.rmdirSync(generatedDir, { recursive: true });
     await generate({
       schemaFile: path.join(schemaDir, "format-schema.js"),
       schemaType: "custom",
@@ -27,25 +26,19 @@ describe("format-schema - compile based", () => {
       "meta.ts",
       "models.ts",
       "schema.json",
-      "validate.ts"
+      "validate.ts",
     ]);
   });
 
   test("schema should match", async () => {
-    const file = fs.readFileSync(
-      path.join(generatedDir, `schema.json`),
-      "utf8"
-    );
+    const file = fs.readFileSync(path.join(generatedDir, `schema.json`), "utf8");
     expect(file).not.toBeUndefined();
     expect(file).toMatchSnapshot();
     expect(await new Ajv().validateSchema(JSON.parse(file))).toEqual(true);
   });
 
   test("decoders should match", () => {
-    const file = fs.readFileSync(
-      path.join(generatedDir, `decoders.ts`),
-      "utf8"
-    );
+    const file = fs.readFileSync(path.join(generatedDir, `decoders.ts`), "utf8");
     expect(file).not.toBeUndefined();
     expect(file).toMatchSnapshot();
   });
@@ -63,22 +56,19 @@ describe("format-schema - compile based - options", () => {
   const schemaDir = path.join(__dirname, "../schemas");
 
   beforeAll(async () => {
-    if (fs.existsSync(generatedDir))
-      fs.rmdirSync(generatedDir, { recursive: true });
+    if (fs.existsSync(generatedDir)) fs.rmdirSync(generatedDir, { recursive: true });
     await generate({
       schemaFile: path.join(schemaDir, "format-schema.js"),
       schemaType: "custom",
       directory: generatedDir,
       addFormats: true,
       formatOptions: { mode: "fast", formats: ["date", "time"] },
+      ajvOptions: { strict: false, allErrors: true },
     });
   });
 
   test("decoders should match", () => {
-    const file = fs.readFileSync(
-      path.join(generatedDir, `decoders.ts`),
-      "utf8"
-    );
+    const file = fs.readFileSync(path.join(generatedDir, `decoders.ts`), "utf8");
     expect(file).not.toBeUndefined();
     expect(file).toMatchSnapshot();
   });
@@ -86,16 +76,11 @@ describe("format-schema - compile based - options", () => {
 
 describe("format-schema - standalone", () => {
   const name = "format";
-  const generatedDir = path.join(
-    __dirname,
-    "../generated",
-    `${name}-standalone`
-  );
+  const generatedDir = path.join(__dirname, "../generated", `${name}-standalone`);
   const schemaDir = path.join(__dirname, "../schemas");
 
   beforeAll(async () => {
-    if (fs.existsSync(generatedDir))
-      fs.rmdirSync(generatedDir, { recursive: true });
+    if (fs.existsSync(generatedDir)) fs.rmdirSync(generatedDir, { recursive: true });
     await generate({
       schemaFile: path.join(schemaDir, "format-schema.js"),
       schemaType: "custom",
@@ -113,16 +98,13 @@ describe("format-schema - standalone", () => {
       "meta.ts",
       "models.ts",
       "schema.json",
-      "validate.ts"
+      "validate.ts",
     ]);
   });
 
   describe("User validator", () => {
     test("validator should contain properties", () => {
-      const file = fs.readFileSync(
-        path.join(generatedDir, `decoders/User/validator.js`),
-        "utf8"
-      );
+      const file = fs.readFileSync(path.join(generatedDir, `decoders/User/validator.js`), "utf8");
       expect(file).not.toBeUndefined();
       expect(file).toContain(`require("ajv-formats/dist/formats").fullFormats.date`);
     });
